[RSDK-11672] Split up ur_arm.cpp into more manageable subfiles
